Moved pow layer to nary_eltwise_layer
