+++ modules/dnn/src/init.cpp
@@ -112,7 +112,6 @@ void initializeLayerFactory()
     CV_DNN_REGISTER_LAYER_CLASS(IsNaN,          IsNaNLayer);
     CV_DNN_REGISTER_LAYER_CLASS(IsInf,          IsInfLayer);
     CV_DNN_REGISTER_LAYER_CLASS(Det,            DetLayer);
-    CV_DNN_REGISTER_LAYER_CLASS(Pow,            PowLayer);
     CV_DNN_REGISTER_LAYER_CLASS(BitShift,       BitShiftLayer);
     CV_DNN_REGISTER_LAYER_CLASS(GridSample,     GridSampleLayer);
