+++ modules/dnn/src/layers/nary_eltwise_layers.cpp
@@ -33,6 +33,41 @@ static int _mod(int x, int y) {
     }
     return res;
 }
+
+// Pow helpers (adapted from legacy PowLayer)
+static void fillBroadcastSteps(const Mat& m, const std::vector<int>& outputShape, std::vector<size_t>& strides)
+{
+    std::vector<int> ms(m.dims);
+    for (int i = 0; i < m.dims; ++i) ms[i] = m.size[i];
+    const size_t dim = std::max(outputShape.size(), ms.size());
+    std::vector<int> msx = ms; msx.insert(msx.begin(), dim - msx.size(), 1);
+    std::vector<size_t> sx(dim, 0);
+    std::vector<size_t> orig(m.step.p, m.step.p + m.dims);
+    std::vector<size_t> origx = orig; origx.insert(origx.begin(), dim - origx.size(), 0);
+    for (size_t i = 0; i < dim; ++i)
+    {
+        if (msx[i] == 1) sx[i] = 0;
+        else sx[i] = origx[i];
+    }
+    strides = sx;
+}
+
+template<typename T>
+static void applyPowSaturate(const double* a, const double* b, Mat& out)
+{
+    T* o = (T*)out.data;
+    const size_t total = out.total();
+    for (size_t i = 0; i < total; i++) o[i] = saturate_cast<T>(std::pow(a[i], b[i]));
+}
+
+static inline void powPlaneFloat(const float* a, const float* b, float* o, int plane_size, size_t dpA, size_t dpB, size_t dpO)
+{
+    if (dpA == 1 && dpB == 1 && dpO == 1) {
+        for (int i = 0; i < plane_size; i++) o[i] = std::pow(a[i], b[i]);
+    } else {
+        for (int i = 0; i < plane_size; i++, a += dpA, b += dpB, o += dpO) *o = std::pow(*a, *b);
+    }
+}
 }
 
 class NaryEltwiseHelper CV_FINAL
@@ -281,7 +316,8 @@ public:
         if (backendId == DNN_BACKEND_CUDA) {
             return op == OPERATION::MAX  || op == OPERATION::MIN  || op == OPERATION::SUM ||
                    op == OPERATION::PROD || op == OPERATION::DIV  || op == OPERATION::ADD ||
-                   op == OPERATION::SUB  || op == OPERATION::MOD || op == OPERATION::FMOD;
+                   op == OPERATION::SUB  || op == OPERATION::MOD || op == OPERATION::FMOD ||
+                   op == OPERATION::POW;
         }
         return backendId == DNN_BACKEND_OPENCV;
     }
@@ -361,12 +397,33 @@ public:
         }
 
         if (op == OPERATION::POW) {
-            /*
-                First input: exponent of Type T;
-                Second input: power of the exponent of Type T1;
-                Output: same type T as first input's.
-            */
-            outputs.assign(1, inputs.front());
+            CV_Assert(inputs.size() == 2);
+            auto isIntegerType = [](int t) {
+                return t == CV_8S || t == CV_8U || t == CV_16S || t == CV_16U || t == CV_32S || t == CV_32U || t == CV_64S || t == CV_64U;
+            };
+            auto isFloatType = [](int t) {
+                return t == CV_32F || t == CV_64F || t == CV_16F || t == CV_16BF;
+            };
+
+            int out_type;
+            const bool baseIsInt   = isIntegerType(inputs[0]);
+            const bool expIsInt    = isIntegerType(inputs[1]);
+            const bool baseIsFloat = isFloatType(inputs[0]);
+            const bool expIsFloat  = isFloatType(inputs[1]);
+
+            if ((baseIsInt && expIsInt) || (baseIsFloat && expIsFloat))
+            {
+                out_type = (inputs[0] == inputs[1]) ? inputs[0] : CV_32F;
+            }
+            else if (baseIsFloat != expIsFloat)
+            {
+                out_type = inputs[0];
+            }
+            else
+            {
+                out_type = CV_32F;
+            }
+            outputs.assign(1, out_type);
             return;
         }
 
@@ -763,6 +820,85 @@ public:
 
         if (inputs.size() == 1) {
             inputs[0].copyTo(outputs[0]);
+            return;
+        }
+
+        if (op == OPERATION::POW) {
+            CV_Assert(inputs.size() == 2);
+            Mat A = inputs[0];
+            Mat B = inputs[1];
+
+            if (A.type() != outputs[0].type()) A.convertTo(A, outputs[0].type());
+            if (B.type() != outputs[0].type()) B.convertTo(B, outputs[0].type());
+
+            CV_Assert(outputs[0].isContinuous());
+            const int nd = outputs[0].dims;
+            std::vector<int> shape(outputs[0].size.p, outputs[0].size.p + nd);
+            std::vector<size_t> step_out(outputs[0].step.p, outputs[0].step.p + nd);
+            std::vector<size_t> step_a(nd, 0), step_b(nd, 0);
+
+            fillBroadcastSteps(A, shape, step_a);
+            fillBroadcastSteps(B, shape, step_b);
+
+            const size_t total = outputs[0].total();
+            size_t dpA = nd ? step_a.back() / outputs[0].elemSize() : 0;
+            size_t dpB = nd ? step_b.back() / outputs[0].elemSize() : 0;
+            size_t dpO = nd ? step_out.back() / outputs[0].elemSize() : 0;
+
+            if (outputs[0].type() == CV_32F) {
+                const float* a = (const float*)A.data;
+                const float* b = (const float*)B.data;
+                float* o = (float*)outputs[0].data;
+                const size_t a_total = A.total();
+                const size_t b_total = B.total();
+                if (a_total == total && b_total == total && dpA == 1 && dpB == 1 && dpO == 1) {
+                    powPlaneFloat(a, b, o, (int)total, 1, 1, 1);
+                } else {
+                    const char* pa0 = (const char*)A.data;
+                    const char* pb0 = (const char*)B.data;
+                    char* po0 = (char*)outputs[0].data;
+                    int plane_size = nd ? shape.back() : 1;
+                    size_t nplanes = 1;
+                    if (nd >= 2) nplanes = std::accumulate(shape.begin(), shape.end()-1, 1, std::multiplies<size_t>());
+                    parallel_for_(Range(0, (int)nplanes), [&](const Range& r){
+                        for (int p = r.start; p < r.end; ++p) {
+                            const char* pa = pa0;
+                            const char* pb = pb0;
+                            char* po = po0;
+                            size_t idx = (size_t)p;
+                            for (int k = nd - 2; k >= 0; k--) {
+                                size_t next_idx = idx / shape[k];
+                                size_t i_k = (int)(idx - next_idx * shape[k]);
+                                pa += i_k * step_a[k];
+                                pb += i_k * step_b[k];
+                                po += i_k * step_out[k];
+                                idx = next_idx;
+                            }
+                            const float* a_ = (const float*)pa;
+                            const float* b_ = (const float*)pb;
+                            float* o_ = (float*)po;
+                            powPlaneFloat(a_, b_, o_, plane_size, dpA, dpB, dpO);
+                        }
+                    });
+                }
+            } else {
+                Mat Ad, Bd; A.convertTo(Ad, CV_64F); B.convertTo(Bd, CV_64F);
+                const double* a = (const double*)Ad.data;
+                const double* b = (const double*)Bd.data;
+                switch (outputs[0].type()) {
+                case CV_32S: applyPowSaturate<int>(a, b, outputs[0]); break;
+                case CV_64S: applyPowSaturate<int64>(a, b, outputs[0]); break;
+                case CV_8U:  applyPowSaturate<uchar>(a, b, outputs[0]); break;
+                case CV_8S:  applyPowSaturate<schar>(a, b, outputs[0]); break;
+                case CV_16U: applyPowSaturate<ushort>(a, b, outputs[0]); break;
+                case CV_16S: applyPowSaturate<short>(a, b, outputs[0]); break;
+                case CV_32U: applyPowSaturate<uint32_t>(a, b, outputs[0]); break;
+                case CV_64U: applyPowSaturate<uint64_t>(a, b, outputs[0]); break;
+                case CV_64F: applyPowSaturate<double>(a, b, outputs[0]); break;
+                default:
+                    CV_Error(Error::StsNotImplemented, "Unsupported integer type for Pow operation");
+                }
+            }
             return;
         }
